ADDING UPDATES

+++ src/main/java/com/osa/model/SalonService.java
@@ -1,13 +1,10 @@
 package com.osa.model;
 
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-
-import com.osa.enums.ServiceName;
+import javax.validation.constraints.NotEmpty;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -22,19 +19,16 @@ public class SalonService {
 	@Id
 	private long serviceId;
 	
-	
-	//@Enumerated(EnumType.STRING)
 	private String serviceName;
 	
-	
+//	@NotEmpty(message="Mention price")
 	private String servicePrice;
 	
-	
+//	@NotEmpty(message = "duration")
 	private String serviceDuration;
 	
-	
-	private int discount;
-	
+//	@NotEmpty(message = "discount")
+	private long discount;	
 //	@OneToOne(mappedBy = "serviceName")
 //	private Appointment appointment;
